Change AddressSpace.accessReg(int, ...) to take a number.
frysk-core/frysk/debuginfo/ChangeLog
2007-11-29  Andrew Cagney  <cagney@redhat.com>

	* DwarfRegisterMapFactory.java: Update; RegisterMap constructor
	takes a String.

frysk-core/frysk/isa/ChangeLog
2007-11-29  Andrew Cagney  <cagney@redhat.com>

	* RegisterMap.java (RegisterMap(String)): New.
	(getRegister(Number)): New.
	(getRegisterNumber(Register)): Check for missing key.
	(getRegister(int)): Ditto.

frysk-core/frysk/stack/ChangeLog
2007-11-29  Andrew Cagney  <cagney@redhat.com>

	* LibunwindRegisterMapFactory.java: Update; RegisterMap
	constructor takes a String.
	* LibunwindAddressSpace.java (accessReg(Number,byte[],boolean)):
	Replace accessFPReg(int,byte[],boolean).
	(getReg(Number)): Replace getReg(int).
	(setReg(Number,long)): Replace setReg(int,long).
	(findRegister(int)): Delete.

frysk-sys/lib/unwind/ChangeLog
2007-11-29  Andrew Cagney  <cagney@redhat.com>

	* AddressSpace.java (getReg(Number)): Replace getReg(int).
	(setReg(Number,long)): Replace setReg(int,long).
	(accessReg(Number,byte[],boolean)): Replace
	accessFPReg(int,byte[],boolean).
	* TestUnwind.java: Update.
	* cni/UnwindX8664.cxx (LIB_UNWIND_REGISTERS_H): Define.
	(TARGET_REGISTERS): Define.
	* cni/UnwindPPC64.cxx: Ditto.
	* cni/UnwindPPC32.cxx: Ditto.
	* cni/UnwindX86.cxx: Ditto.
	* Unwind.java
	(getRegister(RawDataManaged,Number,long,int,byte[],int)): Replace
	getRegister(RawDataManaged,int,long,int,byte[],int).
	(setRegister(RawDataManaged,Number,long,int,byte[],int)): Replace
	setRegister(RawDataManaged,int,long,int,byte[],int).
	* Cursor.java: Update.
	* cni/UnwindH.hxx (access_fpreg): Update.
	(access_reg): Update.
	(getRegister): Update.
	(setRegister): Update.

+++ frysk-sys/lib/unwind/Cursor.java
@@ -71,14 +71,12 @@ public class Cursor
   
     public void getRegister(Number regNum, long offset, int length,
 			    byte[] bytes, int start) {
-	unwinder.getRegister(cursor, regNum.intValue(), offset, length,
-			     bytes, start);
+	unwinder.getRegister(cursor, regNum, offset, length, bytes, start);
     }
 
     public void setRegister(Number regNum, long offset, int length,
 			    byte[] bytes, int start) {
-	unwinder.setRegister(cursor, regNum.intValue(), offset, length,
-			     bytes, start);
+	unwinder.setRegister(cursor, regNum, offset, length, bytes, start);
     }
 
     public long getIP() {
